download tailwind bin
