+++ packages/core/src/lib.rs
@@ -3,6 +3,9 @@
 pub mod schedule;
 pub mod ui_state;
 
+pub mod config {
+    pub use dip_macro::ConfigPlugin;
+}
 pub use dip_task as task;
 
 pub mod prelude {
