+++ src/plugin.rs
@@ -1,0 +1,4 @@
+mod cli;
+mod tool;
+
+pub use crate::plugin::{cli::*, tool::*};
