+++ src/plugin/cli.rs
@@ -1,0 +1,23 @@
+use dip::cli::{CliPlugin, SubcommandPlugin};
+
+#[derive(CliPlugin, clap::Parser)]
+#[clap(version)]
+struct Cli {
+    #[clap(short, long)]
+    path: Option<String>,
+
+    #[clap(subcommand)]
+    action: Action,
+}
+
+#[derive(SubcommandPlugin, clap::Subcommand, Clone, Debug)]
+pub enum Action {
+    #[clap(subcommand)]
+    Tool(ToolAction),
+}
+
+#[derive(SubcommandPlugin, clap::Subcommand, Clone, Debug)]
+pub enum ToolAction {
+    List,
+    Add { name: String },
+}
