+++ src/resource.rs
@@ -1,0 +1,2 @@
+mod platform;
+pub mod tool;
