+++ src/resource/platform.rs
@@ -1,0 +1,15 @@
+pub enum Platform {
+    Linux,
+    Macos,
+    Windows,
+}
+
+impl Platform {
+    pub fn as_str(&self) -> &'static str {
+        match self {
+            Platform::Linux => "linux",
+            Platform::Macos => "macos",
+            Platform::Windows => "windows",
+        }
+    }
+}
